% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariate_drift.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate Inverse Intersection Distance}
\usage{
calculate_distance(variable_old, variable_new, bins = 20)
}
\arguments{
\item{variable_old}{variable from `old` data}

\item{variable_new}{variable from `new` data}

\item{bins}{continuous variables are discretized to `bins` intervals of equal size}
}
\value{
inverse intersection distance
}
\description{
Calculate Inverse Intersection Distance
}
