% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{centroid_colors}
\alias{centroid_colors}
\title{Function to assign colors to centroid points}
\usage{
centroid_colors(x)
}
\arguments{
\item{x}{centroid points sf object}
}
\value{
centroid points sf object with colors
}
\description{
Function to assign colors to centroid points
}
