% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{centroid_distance}
\alias{centroid_distance}
\title{Function to calculate distance between start and end points}
\usage{
centroid_distance(x)
}
\arguments{
\item{x}{centroid points df}
}
\value{
distance between start and end points
}
\description{
Function to calculate distance between start and end points
}
