% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{centroid_labels}
\alias{centroid_labels}
\title{Function to add an A and B label to first and last point of centroid line}
\usage{
centroid_labels(x)
}
\arguments{
\item{x}{centroid points sf object}
}
\value{
centroid points sf object with A and B label on first and last points
}
\description{
Function to add an A and B label to first and last point of centroid line
}
