% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{centroid_sizes}
\alias{centroid_sizes}
\title{Function to assign a size to centroid points}
\usage{
centroid_sizes(x)
}
\arguments{
\item{x}{centroid points sf object}
}
\value{
centroid points sf object with point sizes
}
\description{
Function to assign a size to centroid points
}
