% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pDrift.R
\name{check.pDrift}
\alias{check.pDrift}
\title{\code{check.pDrift} Visual check of drift rate and weight calculations}
\usage{
check.pDrift(data, ptMag = 2)
}
\arguments{
\item{data}{Data frame with dive data output from the \code{pDrift} function.}

\item{ptMag}{Scaling of point sizes to increase or decrease emphasis}
}
\value{
Interactive plot
}
\description{
\code{check.pDrift} Visual check of drift rate and weight calculations
}
\details{
Opens an interactive plot of entire dive record, where dives can be individually
checked by selecting them. Opens a second plot with the individual dive profile.
}
\examples{
rbs(data=dive, num=100, n.bs=4)
dive <- rbs(data=dive, num=NA, n.bs=4)
dive <- pDrift(dive)
check.pDrift(dive)
}
\seealso{
\code{\link{rbs}} Reverse Broken Stick algorithm

\code{\link{pDrift}} to calculate drift rate and probability weightings for drift dive detection

\code{\link{fitDrift}} To fit state-space model to weighted data to estimate drift rate change over time

\code{\link{fitPlot}} To plot output from \code{fitDrift}

Other Drift dive functions: 
\code{\link{fitDrate}()},
\code{\link{fitPlot}()},
\code{\link{pDrift}()},
\code{\link{rbs}()}
}
\author{
Martin Biuw
}
\concept{Drift dive functions}
