% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_model_drift.R
\name{compare_two_profiles}
\alias{compare_two_profiles}
\title{Calculates distance between two Ceteris Paribus Profiles}
\usage{
compare_two_profiles(cpprofile_old, cpprofile_new, variables, scale = 1)
}
\arguments{
\item{cpprofile_old}{Ceteris Paribus Profile for historical / `old` model}

\item{cpprofile_new}{Ceteris Paribus Profile for current / `new` model}

\item{variables}{variables for which drift should be calculated}

\item{scale}{scale parameter for calculation of scaled drift}
}
\value{
data frame with distances between Ceteris Paribus Profiles
}
\description{
This function calculates square root from mean square difference between Ceteris Paribus Profiles
}
