% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelp_farms.R
\name{create_kelp_farm_polygons}
\alias{create_kelp_farm_polygons}
\title{Function to create polygons out of the kelp farm coords above}
\usage{
create_kelp_farm_polygons(kelp_farms_coords, crs = 4326)
}
\arguments{
\item{crs}{coordinate reference system}

\item{x}{list of kelp farms and their coordinates}
}
\value{
list of kelp farm polygons
}
\description{
Function to create polygons out of the kelp farm coords above
}
