% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{detect_filetype}
\alias{detect_filetype}
\title{Function to determine the filetype given a filename}
\usage{
detect_filetype(filename)
}
\arguments{
\item{filename}{name of the file}
}
\value{
chr
}
\description{
Function to determine the filetype given a filename
}
