% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{drift_area}
\alias{drift_area}
\title{Function that given a GPX file of track points, will output the convex hull polygon of interest}
\usage{
drift_area(x, d)
}
\arguments{
\item{x}{sf, drift data}

\item{d}{num, buffer distance in meters around the track}
}
\value{
sf, a convex hull polygon object
}
\description{
Function that given a GPX file of track points, will output the convex hull polygon of interest
}
