% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{drift_speed}
\alias{drift_speed}
\title{Function to calculate drift speed given a table of drifter points}
\usage{
drift_speed(x)
}
\arguments{
\item{x}{tbl of drift data that has point locations and time}
}
\value{
tbl of drift data but with a new column of calculated speed (m/s) and great circle distance
}
\description{
Function to calculate drift speed given a table of drifter points
}
