% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drifter-package.R
\docType{package}
\name{drifter}
\alias{drifter-package}
\alias{drifter}
\title{drifter package}
\description{
Supplementary tools working with drifter GPX data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kolive4/drifter}
}

}
