% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{elapsed_time}
\alias{elapsed_time}
\title{Function to produce the average elapsed time}
\usage{
elapsed_time(x)
}
\arguments{
\item{x}{sf object, drifts with times}
}
\value{
chr, string of the average elapsed time across three drifters
}
\description{
Function to produce the average elapsed time
}
