% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.R, R/io.R
\name{extract_tracks}
\alias{extract_tracks}
\title{Extract one or more tracks}
\usage{
extract_tracks(x)

extract_tracks(x)
}
\arguments{
\item{x}{xml_nodeset of one or more "trk" nodes}
}
\value{
a table with Name, Longitude, Latitude, Elevation and Time

a table with Name, Longitude, Latitude, Elevation and Time
}
\description{
Extract one or more tracks

Extract one or more tracks
}
