% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.R, R/io.R
\name{extract_trksegs}
\alias{extract_trksegs}
\title{Extract a track segments}
\usage{
extract_trksegs(trkseg)

extract_trksegs(trkseg)
}
\arguments{
\item{trkseg}{xmlnode_set of one or more "trkpt" elements}
}
\value{
a character matrix with unnamed columns Longitude, Latitude, Elevation and Time

a character matrix with unnamed columns Longitude, Latitude, Elevation and Time
}
\description{
Extract a track segments

Extract a track segments
}
