% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDrate.R
\name{fitDrate}
\alias{fitDrate}
\title{\code{fitDrate} Fit state-space model to weighted drift dive record}
\usage{
fitDrate(data, weight = "default", prefilter = 0)
}
\arguments{
\item{data}{Data frame with dive data output from the \code{pDrift} function.}

\item{weight}{Select weightingbvariables to be used. Default is to use \code{weight} created by \code{pDrift}.
See example for how to specify other weighting}

\item{prefilter}{Should any prefiltering be done (based on selected weight) prior to running the model.}
}
\value{
A list with several objects from fit:
\item{\code{data}}{Original input data}
\item{\code{w}}{Weighting used for probability of drift dive}
\item{\code{tSeq}}{Regular time series (e.g. whole days) where predictions are obtained}
\item{\code{obj}}{TMB object}
\item{\code{opt}}{TMB optimization}
\item{\code{par}}{Fitted parameters}
\item{\code{rep}}{TMB report}
\item{\code{allsd}}{Standard deviation of fitted values}
\item{\code{plsd}}{Fitted values}
}
\description{
\code{fitDrate} Fit state-space model to weighted drift dive record
}
\details{
Takes a dive set of dive data from SRDLs (Sea Mammal Research Unit), with added variables from
\code{rbs} and \code{pDrift} and fits a state-space model to change over time (using TMB).
}
\examples{
rbs(data=dive, num=100, n.bs=4)
dive <- rbs(data=dive, num=NA, n.bs=4)
dive <- pDrift(dive)
fit <- fitDrate(dive, prefilter=0.01)
fit2 <- fitDrate(dive, weight='rRes*span*dWt', prefilter=0.01)
}
\seealso{
\code{\link{rbs}} Reverse Broken Stick algorithm

\code{\link{check.pDrift}} to interactively check drift dive weightings

\code{\link{pDrift}} to calculate drift rate and probability weightings for drift dive detection

\code{\link{fitPlot}} To plot output from \code{fitDrift}

Other Drift dive functions: 
\code{\link{check.pDrift}()},
\code{\link{fitPlot}()},
\code{\link{pDrift}()},
\code{\link{rbs}()}
}
\author{
Martin Biuw
}
\concept{Drift dive functions}
