% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{geom_coastline}
\alias{geom_coastline}
\title{Function for mapping the coastline}
\usage{
geom_coastline(
  coast = rnaturalearth::ne_coastline(scale = "small", returnclass = "sf"),
  bb = NULL,
  ...
)
}
\arguments{
\item{coast}{coastline object from rnaturalearth}

\item{bb}{bounding box coordinates for cropping}

\item{...}{arguments passable to geom_sf}
}
\value{
x, ggplot layer
}
\description{
Function for mapping the coastline
}
