% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{get_bb_table}
\alias{get_bb_table}
\title{Function to create a bounding box table}
\usage{
get_bb_table(
  name = c("casco", "chcv", "stgeo", "mecoast", "cheb"),
  xmin = c(-70.288, -70.167931, -69.301235, -70.288, -70.16),
  xmax = c(-69.715072, -70.102219, -69.246169, -69.016407, -70.121795),
  ymin = c(43.549256, 43.701498, 43.921795, 43.549256, 43.71),
  ymax = c(43.921978, 43.723947, 43.9558, 44.137128, 43.740928),
  long_name = c("TBI-Casco Bay", "TBI-Chandler Cove", "TBI-St. George River",
    "ME coast to St. George", "West side of Chebeague Island")
)
}
\arguments{
\item{name}{chr, abbreviated name of bb location}

\item{xmin}{num, min longitude}

\item{xmax}{num, max longitude}

\item{ymin}{num, min latitude}

\item{ymax}{num, max latitude}

\item{long_name}{chr, long name of bb location}
}
\value{
bounding box table
}
\description{
Function to create a bounding box table
}
