% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{get_bbox}
\alias{get_bbox}
\title{Function that takes one element of the table and returns a bbox}
\usage{
get_bbox(n)
}
\arguments{
\item{n}{name of bounding box}
}
\value{
a bounding box object to be used with st_crop or other manipulations
}
\description{
Function that takes one element of the table and returns a bbox
}
