% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.R
\name{gpx.extract_tracks}
\alias{gpx.extract_tracks}
\title{Extract one or more tracks}
\usage{
gpx.extract_tracks(x)
}
\arguments{
\item{x}{xml_nodeset of one or more "trk" nodes}
}
\value{
a table with Name, Longitude, Latitude, Elevation and Time
}
\description{
Extract one or more tracks
}
