% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.R
\name{gpx.extract_trksegs}
\alias{gpx.extract_trksegs}
\title{Extract a track segments}
\usage{
gpx.extract_trksegs(trkseg)
}
\arguments{
\item{trkseg}{xmlnode_set of one or more "trkpt" elements}
}
\value{
a character matrix with unnamed columns Longitude, Latitude, Elevation and Time
}
\description{
Extract a track segments
}
