% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{gpx_to_sf}
\alias{gpx_to_sf}
\title{Function to convert the gpx track to an sf object for plotting}
\usage{
gpx_to_sf(trax_tibble)
}
\arguments{
\item{trax_tibble}{gpx object as a tibble}
}
\value{
sf object of gpx track
}
\description{
Function to convert the gpx track to an sf object for plotting
}
