% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{gpx_to_tibble}
\alias{gpx_to_tibble}
\title{Function to convert gpx track to tibble for manipulation}
\usage{
gpx_to_tibble(gpx)
}
\arguments{
\item{gpx}{gpx object}
}
\value{
tibble version of gpx
}
\description{
Function to convert gpx track to tibble for manipulation
}
