% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{import_csv}
\alias{import_csv}
\title{Function to import CSV with tracks/point locations along a track}
\usage{
import_csv(
  filename,
  crs = 4326,
  coords = c("longitude", "latitude", "date_time"),
  tz = "UTC",
  ...
)
}
\arguments{
\item{filename}{name of file}

\item{crs}{coordinate reference system}

\item{coords}{names of coordinates and date_time columns}

\item{tz}{chr, time zone}

\item{...}{who knows}
}
\value{
a table or sf object
}
\description{
Function to import CSV with tracks/point locations along a track
}
