% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_date_time}
\alias{import_date_time}
\title{function to deal with time column formatting}
\usage{
import_date_time(x, coords, tz = "UTC")
}
\arguments{
\item{x}{sf table}

\item{coords}{date_time coordinate column name}

\item{tz}{time zone}
}
\value{
updated table if error not thrown, with a date_time_ column
}
\description{
function to deal with time column formatting
}
