% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_tracks}
\alias{import_tracks}
\title{Import any track object as an sf object}
\usage{
import_tracks(
  filename = system.file("ex_data/cape_cod_complicated.GPX", package = "drifter"),
  type = detect_filetype(filename),
  ...
)
}
\arguments{
\item{filename}{name of track file}

\item{type}{chr, declares what the data format is (gpx, csv, ...)}

\item{...}{additional params to import_*}
}
\description{
Import any track object as an sf object
}
