% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelp_farms.R
\name{kelp_farms_coords}
\alias{kelp_farms_coords}
\title{Function to return polygons of kelp farms}
\usage{
kelp_farms_coords()
}
\value{
list of kelp farm polygon matrices
}
\description{
Function to return polygons of kelp farms
}
