% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{leaflet_drifter}
\alias{leaflet_drifter}
\title{Draw a drifter as a leaflet polyline}
\usage{
leaflet_drifter(x)
}
\arguments{
\item{x}{sf object, drifter track}
}
\value{
leaflet map with drifter track
}
\description{
Draw a drifter as a leaflet polyline
}
