% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{particle_depth}
\alias{particle_depth}
\title{Function that when given a polygon with an averaged speed and direction and a sinking rate, shows what depth kelp particles are at at a given distance within a polygon}
\usage{
particle_depth(area, sink_rate, start_depth, max_depth, farms)
}
\arguments{
\item{area}{sf polygon, area polygon of interest}

\item{sink_rate}{num, rate of sinking in m/s}

\item{start_depth}{num, kelp planting depth in m}

\item{max_depth}{num, static depth in m}

\item{farms}{sf polygon, kelp farm}
}
\value{
stars object with particle depths
}
\description{
Function that when given a polygon with an averaged speed and direction and a sinking rate, shows what depth kelp particles are at at a given distance within a polygon
}
