% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pDrift.R
\name{pDrift}
\alias{pDrift}
\title{\code{pDrift} Drift rate and weight calculations}
\usage{
pDrift(data = dive, depth.wpar = c(100, 30), plotit = F, keep.all = T)
}
\arguments{
\item{data}{Data frame with dive data output from the \code{rbs} function.}

\item{depth.wpar}{Set mean ad standard deviation of half gaussian for depth weighting}

\item{plotit}{Logical, plot results or not}

\item{keep.all}{Logical, keep all data or only data containing valid drift rates and weights}
}
\value{

}
\description{
\code{pDrift} Drift rate and weight calculations
}
\details{
Takes a dive set of dive data from SRDLs (Sea Mammal Research Unit), and calculates
drift rate based on segment between first two extracted inflection points, along with
probability weights for each dive being a drift dive.
}
\examples{
rbs(data=dive, num=100, n.bs=4)
dive <- rbs(data=dive, num=NA, n.bs=4)
dive <- pDrift(dive)
}
\seealso{
\code{\link{rbs}} Reverse Broken Stick algorithm

\code{\link{check.pDrift}} to interactively check drift dive weightings

\code{\link{fitDrate}} To fit state-space model to weighted data to estimate drift rate change over time

\code{\link{fitPlot}} To plot output from \code{fitDrift}

Other Drift dive functions: 
\code{\link{check.pDrift}()},
\code{\link{fitDrate}()},
\code{\link{fitPlot}()},
\code{\link{rbs}()}
}
\author{
Martin Biuw
}
\concept{Drift dive functions}
