% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariate_drift.R
\name{print.covariate_drift}
\alias{print.covariate_drift}
\title{Print Covariate Drift Data Frame}
\usage{
\method{print}{covariate_drift}(x, max_length = 25, ...)
}
\arguments{
\item{x}{an object of the class `covariate_drift`}

\item{max_length}{length of the fist column, by default 25}

\item{...}{other arguments, currently ignored}
}
\value{
this function prints a data frame with a nicer format
}
\description{
Print Covariate Drift Data Frame
}
\examples{
library("DALEX2")
# here we do not have any drift
d <- calculate_covariate_drift(apartments, apartments_test)
d
# here we do have drift
d <- calculate_covariate_drift(dragons, dragons_test)
d

}
