% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_model_drift.R
\name{print.model_drift}
\alias{print.model_drift}
\title{Print Model Drift Data Frame}
\usage{
\method{print}{model_drift}(x, max_length = 25, ...)
}
\arguments{
\item{x}{an object of the class `model_drift`}

\item{max_length}{length of the first column, by default 25}

\item{...}{other arguments, currently ignored}
}
\value{
this function prints a data frame with a nicer format
}
\description{
Print Model Drift Data Frame
}
\examples{
 library("DALEX2")
 \dontrun{
 library("ranger")
 predict_function <- function(m,x,...) predict(m, x, ...)$predictions
 model_old <- ranger(m2.price ~ ., data = apartments)
 model_new <- ranger(m2.price ~ ., data = apartments_test)
 calculate_model_drift(model_old, model_new,
                  apartments_test,
                  apartments_test$m2.price,
                  max_obs = 1000,
                  predict_function = predict_function)

 predict_function <- function(m,x,...) predict(m, x, ..., probability=TRUE)$predictions[,1]
 data_old = HR[HR$gender == "male", -1]
 data_new = HR[HR$gender == "female", -1]
 model_old <- ranger(status ~ ., data = data_old, probability=TRUE)
 model_new <- ranger(status ~ ., data = data_new, probability=TRUE)
 calculate_model_drift(model_old, model_new,
                  HR_test,
                  HR_test$status == "fired",
                  max_obs = 1000,
                  predict_function = predict_function)

 # plot it
 library("ceterisParibus2")
 prof_old <- individual_variable_profile(model_old,
                                     data = data_new,
                                     new_observation = data_new[1:1000,],
                                     label = "model_old",
                                     predict_function = predict_function)
 prof_new <- individual_variable_profile(model_new,
                                     data = data_new,
                                     new_observation = data_new[1:1000,],
                                     label = "model_new",
                                     predict_function = predict_function)
 plot(prof_old, prof_new,
      selected_variables = "age", aggregate_profiles = mean,
      show_observations = FALSE, color = "_label_")

}

}
