% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbs.R
\name{rbs}
\alias{rbs}
\title{\code{rbs} Reverse Broken Stick algorithm}
\usage{
rbs(data = dive, num = 100, n.bs = 4)
}
\arguments{
\item{data}{Data frame with dive data output from Satellite Relay Data Loggerds (SRDLs)}

\item{num}{Which dive in data should be processed. If NA, the entire dataset is processed.}

\item{n.bs}{Number of valid broken stick time/depth points in data.}
}
\description{
\code{rbs} Reverse Broken Stick algorithm
}
\details{
Takes a standard dive set of dive data from SRDLs (Sea Mammal Research Unit),
and returns three new variables: 
\code{order}: Character, order in which inflction points were extracted
\code{max.res}: Numeric, maximum remaining residual (in meters) unaccounted for by summarized dive profile
\code{reordered}: Logical, if dive was reorganized from inconsistent ordering in original file    
NOTE! Names in the data frame sould have underscores replaced by dots prior to running this function!
}
\examples{
rbs(data=dive, num=100, n.bs=4)
dive <- rbs(data=dive, num=NA, n.bs=4)

}
\seealso{
\code{\link{pDrift}} to calculate drift rate and probability weightings for drift dive detection

\code{\link{check.pDrift}} to interactively check drift dive weightings

\code{\link{fitDrift}} To fit state-space model to weighted data to estimate drift rate change over time

\code{\link{fitPlot}} To plot output from \code{fitDrift}

Other Drift dive functions: 
\code{\link{check.pDrift}()},
\code{\link{fitDrate}()},
\code{\link{fitPlot}()},
\code{\link{pDrift}()}
}
\author{
Martin Biuw
}
\concept{Drift dive functions}
