% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_gpx}
\alias{read_gpx}
\title{Funciton to read GPX files}
\usage{
read_gpx(
  filename = list.files(system.file("ex_data", package = "drifter"), full.names = TRUE),
  form = c("tibble", "sf")[1],
  which = c("routes", "tracks", "waypoints")[2]
)
}
\arguments{
\item{filename}{filename(s) of GPX file}

\item{form}{output format (sf, tbl)}

\item{which}{one of (routes, tracks, waypoints)}
}
\value{
either a tibble or sf object of a GPX object
}
\description{
Funciton to read GPX files
}
