% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide2time.R
\name{tide.find_closest_station}
\alias{tide.find_closest_station}
\title{Function to match US tide stations to drift centroids}
\usage{
tide.find_closest_station(x = import_tracks(), n = 10, grouping_id = "Name")
}
\arguments{
\item{x}{sf table of one (or more) drift track(s)}

\item{n}{number of closest stations to be returned}

\item{grouping_id}{if there are multiple tracks, which column defines the grouping variable}
}
\value{
the maree tide station table, sorted from closest to furthest from average drift track lat/long
}
\description{
Function to match US tide stations to drift centroids
}
