% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{trial_centroid}
\alias{trial_centroid}
\title{Function to find centroid of drifts}
\usage{
trial_centroid(df = import_tracks())
}
\arguments{
\item{df}{dataframe of gpx points of drifters for one trial}
}
\value{
sf centroid linestring
}
\description{
Function to find centroid of drifts
}
