% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariate_drift.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate Non-Intersection Distance}
\usage{
calculate_distance(variable_old, variable_new, bins = 20)
}
\arguments{
\item{variable_old}{variable from `old` data}

\item{variable_new}{variable from `new` data}

\item{bins}{continuous variables are discretized to `bins` intervals of equal size}
}
\value{
Non-Intersection Distance
}
\description{
Calculate Non-Intersection Distance
}
\examples{
calculate_distance(rnorm(1000), rnorm(1000))
calculate_distance(rnorm(1000), runif(1000))
}
